feature/0500: commit перед сменой репозитория

+++ services/account-service/src/main/java/com/pot/app/accountservice/service/AccountReservationService.java
@@ -3,11 +3,12 @@ package com.pot.app.accountservice.service;
 import com.pot.app.accountservice.dto.BalanceData;
 import com.pot.app.accountservice.dto.ReservationResult;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface AccountReservationService {
 
-    ReservationResult reserve(String transactionId, String userId, long amount);
+    ReservationResult reserve(String transactionId, String userId, BigDecimal amount);
 
     boolean commit(UUID reservationId, String transactionId);
